better handling of admin login
django admin overrides the normal login mechanism, not giving our view a chance
to stick the final destination URL in the session. CAS requires that
when we validate the ticket, we also send along that final destination URL
that we originally specified.

So, in our caslogin view, if we don't see a cached URL in the session,
we take our best guess and assume that it was the admin login that
got us into this situation and try to generate the equivalent URL
that would match that scenario.