let's actually pull out the affils
TR uses the 'cas:courses' element. There doesn't seem
to be any real standard here, so the most straightforward
course of action for now is to just look for those
explicitly as well. It should be a rare enough occasion
that we have to support a different provider's format
that hardcoding these elements should be ok.