Use '' for default CAS_DEFAULT_NEXT setting
This resolves an issue with new CAS server. The 'next' param
needs to be the same as in the CAS input form. I'm not sure why this
weird param was added as a default here, but I'm able to log in on CAS
on both the front page, and the admin form, on mediathread-staging with
this change.

As Dan mentioned, we eventually want to clean this up (change
'destination' to 'TARGET' on the log in form). Making that change broke
the login for me, so I'm keeping that as is at the moment. That's in
mediathread_deploy_specific, btw.